refactor: move JS files
