+++ invenio_jobs/assets/semantic-ui/js/invenio_jobs/administration/StopButton.js
